requirements.txt
google-cloud-datastore
google-cloud-storage
google-cloud-bigquery